+++ tests/smoke/CSharp80/NullableReferenceTypes/EnableNullableContextAndDeclareFieldAsNullable/CommonClassWithNullableField.cs
@@ -1,6 +1,6 @@
 ﻿// Expected number of suggestions: 1
 
-namespace CSharp80.NullableReferenceTypes.EnableNullableContextAndDeclareReferenceFieldAsNullable
+namespace CSharp80.NullableReferenceTypes.EnableNullableContextAndDeclareFieldAsNullable
 {
     public class CommonClassWithNullableField
     {